+++ src/router/index.js
@@ -134,13 +134,6 @@ const router = new Router({
 });
 
 router.beforeEach(async (to, from, next) => {
-  if (
-    to.path.startsWith("/external-transferred") &&
-    !getNetworkItemByAppId(to.path.split("-").slice(-1)[0]).isEnabled
-  ) {
-    next("/404");
-  }
-
   await store.dispatch("onRouteChange", from);
   await store.dispatch("onFolderChange", to.query.id);