+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [v3.5.4] - 2021-05-09
+- hotfix: ([85](https://github.com/meateam/drive-ui/pull/85)): add validation to search by mail or t.
+- hotfix: ([85](https://github.com/meateam/drive-ui/pull/85)): show all external transffers in the sidenav.
+- hotfix: ([81](https://github.com/meateam/drive-ui/pull/81)): make approverInfo request async.
+
 ## [v3.5.3] - 2021-05-02
 - hotfix: ([81](https://github.com/meateam/drive-ui/pull/81)): decrease approverInfo request timeout.
 - hotfix: ([80](https://github.com/meateam/drive-ui/pull/80)): fix networkDest.js read property of undefined when approverInfo request fails.
@@ -101,6 +106,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - default timoout for the axios request because some big files uploads that took more then 10 sec faild.
 
 [unreleased]: https://github.com/meateam/api-gateway/compare/master...develop
+[v3.5.4]: https://github.com/meateam/api-gateway/compare/v3.5.3...v3.5.4
 [v3.5.3]: https://github.com/meateam/api-gateway/compare/v3.5.2...v3.5.3
 [v3.5.2]: https://github.com/meateam/api-gateway/compare/v3.5.1...v3.5.2
 [v3.5.1]: https://github.com/meateam/api-gateway/compare/v3.4.0...v3.5.1